s2a/src/backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-dotenv>=1.0.0
python-multipart>=0.0.12
pandas>=2.2.0
numpy>=1.26.0
pydantic>=2.9.0
requests>=2.32.0
beautifulsoup4>=4.12.0
openai>=1.50.0
sqlalchemy>=2.0.0
scipy>=1.12.0
scikit-learn>=1.4.0
